feat: Implement load balancing with circuit breaker policy
This commit modifies the APIM policies module to implement load balancing with a circuit breaker policy. It introduces the `lb_pool` and `backends` directories in the Terraform module, which contain the necessary code for configuring the LB pool and backends. The `apimanagement.tf` file is updated to use the new `lb_pool` module and pass the required parameters. The `lb-pool.tf` file, previously located in the `load-balancing` directory, is deleted as it is no longer needed.